pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-linter"
version = "0.1.0"
description = "A validation tool for AI skills and agent configurations"
readme = "README.md"
requires-python = ">=3.10"
license="MIT"
license-files = ["LICENSE"]
authors = [
    {name = "François Chastanet"}
]
keywords = ["ai", "linter", "validation", "skills", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "tiktoken>=0.5.0",
    "PyYAML>=6.0",
    "pyparsing>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
    "build>=0.8.0",
    "twine>=4.0.0"
]

[project.urls]
Homepage = "https://github.com/fchastanet/ai-linter"
Repository = "https://github.com/fchastanet/ai-linter.git"
Issues = "https://github.com/fchastanet/ai-linter/issues"

[project.scripts]
ai-linter = "aiLinter:main"

[tool.setuptools]
py-modules = ["aiLinter"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
install_types = true
non_interactive = true
# Point mypy to your source root
mypy_path = "src"
# Disallow imports as src.lib.*
namespace_packages = true
# Show error codes
show_error_codes = true
# Ignore missing imports from external packages
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src --cov-report=html --cov-report=term-missing"
testpaths = ["tests"]
